refactor: Creating http package to hold networking services
